pyproject.toml
[tool.poetry]
name = "pythonbankingoperations"
version = "0.1.0"
description = ""
authors = ["Dmitriy Koval"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
flake8 = "^7.1.1"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"


[tool.poetry.group.lint.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = ".git"

[tool.isort]
# максимальная длина строки
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'